Relocate config files for the mobile
