--- apps/mobile/.watchmanconfig
@@ -1,1 +1,0 @@
-{}
