--- apps/mobile/babel.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  presets: ['module:metro-react-native-babel-preset'],
-};
