+++ apps/mobile/src/__tests__/App-test.tsx
