+++ tsconfig.json
@@ -1,35 +1,20 @@
 {
   "compilerOptions": {
-    "allowJs": false,
-    "noImplicitReturns": true,
-    "noImplicitThis": true,
-    "noImplicitAny": false,
-    "strictNullChecks": true,
+    "allowJs": true,
+    "allowSyntheticDefaultImports": true,
     "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
     "isolatedModules": true,
-    "jsx": "preserve",
-    "lib": ["dom", "es2017", "dom.iterable"],
-    "module": "esnext",
+    "jsx": "react",
+    "lib": ["es6"],
     "moduleResolution": "node",
     "noEmit": true,
-    "noFallthroughCasesInSwitch": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "resolveJsonModule": true,
-    "skipLibCheck": true,
     "strict": true,
-    "target": "esnext",
-    "experimentalDecorators": true,
-    "typeRoots": ["node_modules/@types", "types", "src/types"],
-    "sourceMap": true,
-    "sourceRoot": "src",
-    "baseUrl": ".",
-    "paths": {
-      "src/*": ["src/*"]
-    },
-    "plugins": [{"name": "typescript-tslint-plugin"}]
+    "target": "esnext"
   },
-  "exclude": ["node_modules"],
-  "include": ["**/*.d.ts", "**/*.ts", "**/*.tsx"]
+  "exclude": [
+    "node_modules",
+    "babel.config.js",
+    "metro.config.js",
+    "jest.config.js"
+  ]
 }
